requirements.txt
asyncpg==0.30.0
aiogram==3..17.0
psycopg2-binary==2.9.10
APScheduler=3.11.0
python-decouple==3.8
python-dotenv=1.0.1